キャストを追加。
git-svn-id: svn://localhost/kirikiri2/trunk@855 131d44d2-bf85-024a-9602-49a0ea93a7a9

+++ kirikiri2/src/core/visual/win32/krmovie/BufferRenderer.cpp
@@ -628,7 +628,7 @@ void TBufferRenderer::OnRenderStart( IMediaSample *pMediaSample )
 	{
 		TimeStart = m_StartFrame + Current;
 		TimeEnd = m_StartFrame + Current;
-		m_StopFrame = m_StartFrame + Stop;
+		m_StopFrame = m_StartFrame + static_cast<LONG>(Stop);
 	}
 	else
 	{
